fireflies-meeting-analyzer/requirements.txt
# Fireflies Meeting Analyzer Dependencies

# Web Framework & Dashboard
streamlit>=1.31.0
plotly>=5.18.0

# HTTP & API
requests>=2.31.0
httpx>=0.26.0

# Data Processing
pandas>=2.2.0
python-dateutil>=2.8.2

# AI/LLM Integration
anthropic>=0.18.0
openai>=1.12.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Export & Reporting
openpyxl>=3.1.2
markdown>=3.5.0
jinja2>=3.1.3

# Utilities
rich>=13.7.0
typer>=0.9.0
